Add files via upload
The action attribute specifies the URL where the form data will be sent when the form is submitted.
The method attribute specifies the HTTP method to be used (commonly "get" or "post").
The form contains input fields for a username and password, and a submit button.
Forms play a crucial role in enabling user interactions on websites, allowing users to input information, make selections, and trigger actions. After submission, the data is typically processed on the server side using server-side scripting languages such as PHP, Python, or Node.js.